Refactored top level components from js to react
